# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for EasyFind application.",
        "title": "EasyFind API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/user/account": {
            "delete": {
                "description": "硬删除账号。普通用户只能注销自己，管理员/系统管理员可以注销任意账号。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "注销/删除账号 (硬删除)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "目标用户名 (管理员使用)",
                        "name": "username",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "注销成功",
                        "schema": {
                            "$ref": "#/definitions/easyfind_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/user/profile": {
            "get": {
                "description": "根据 Username 获取用户资料",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "获取个人资料",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "username",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/easyfind_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/easyfind_internal_models.Account"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "修改当前登录用户的资料，为空的字段不修改",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "修改个人资料",
                "parameters": [
                    {
                        "description": "需要修改的参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controllers.UpdateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改成功",
                        "schema": {
                            "$ref": "#/definitions/easyfind_pkg_response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "easyfind_internal_models.Account": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "first_login": {
                    "description": "是否首次登录",
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "nickname": {
                    "description": "昵称",
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "description": "1:学生/老师, 2:失物招领管理员, 3:系统管理员",
                    "allOf": [
                        {
                            "$ref": "#/definitions/easyfind_internal_models.UserRole"
                        }
                    ]
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "description": "学号或工号",
                    "type": "string"
                }
            }
        },
...
        "easyfind_internal_models.UserRole": {
            "type": "integer",
            "enum": [
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "RoleStudentTeacher",
                "RoleLFAdmin",
                "RoleSysAdmin"
            ]
        },
        "easyfind_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "description": "使用 data 统一包裹数据"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "internal_controllers.UpdateProfileRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}